- Added EventHandlers directory - Moved some logic into those files - Fixed some Types in the Actions file
